+++ src/assets/styles/_page-layout.scss
@@ -10,6 +10,10 @@
 	}
 }
 
+.ant-table-wrapper tbody {
+	background: #ffffff;
+}
+
 .content-wrapper-simplified {
 	transition: all 0.3s;
 }
