+++ src/pages/admin-panel/investments/Block.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { connect } from "react-redux";
 import { Formik } from "formik";
 import { Card, Button, Input, Form, Typography } from "antd";
-import { Block } from "../../../redux/admin-panel/investments";
+import { Block } from "../../../redux/admin-panel/investment";
 import { createSecret } from "./../../../utils/secretHash";
 import { TextAligner } from "./../../../components/styled";
 const { Title } = Typography;
